requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
pydantic==2.11.3
python-dotenv==1.1.0
yfinance==0.2.55
openai==1.75.0
requests
